% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boot.param.sd}
\alias{boot.param.sd}
\title{Calculate the standard deviation of one or more parameters by bootstraping}
\usage{
boot.param.sd(x, nboot = 1000)
}
\arguments{
\item{x}{a vector or matrix (with each column as one parameter)}

\item{nboot}{number of bootstraps; default: 1000}
}
\value{
the sd of a vector, or a data.frame with the sd of each parameter
in each row
}
\description{
Calculate the standard deviation of one or more parameters by bootstraping
}
\author{
Nikolaos Tourvas
}
