% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genemapper2df.R
\name{genemapper2df}
\alias{genemapper2df}
\title{Combine txt files generated by GeneMapper v4.1}
\usage{
genemapper2df(folder_name, export_name = "genemapper_export.csv",
  export_option = "csv")
}
\arguments{
\item{folder_name}{a folder where all the .txt files
generated by GeneMapper are stored.}

\item{export_name}{the name of the exported csv file.}

\item{export_option}{whether to export a csv file (option "csv")
or a data.frame in the R Environment (option "R").}
}
\value{
a single csv file or a data.frame
}
\description{
Combine txt files generated by GeneMapper v4.1
}
\examples{

# Export csv file for use in GenAlEx
genemapper2df(folder_name = "...",
              export_name = "my_data",
              export_option = "csv")

# Create genind object
df <- genemapper2df(folder_name = "...", export_option = "R")
obj <- loci2genind(as.loci(df))

}
