% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemix.R
\name{genind2treemix}
\alias{genind2treemix}
\title{Generate microsatellite input file for TreeMix}
\usage{
genind2treemix(x, export_name = ".")
}
\arguments{
\item{x}{a genind object}

\item{export_name}{Name of the exported input file. Note that it should include the path to the directory you would like to export it to.}
}
\value{
the input file for TreeMix as is required for microsatellite data
}
\description{
Generate microsatellite input file for TreeMix
}
\author{
Nikolaos Tourvas
}
