% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pic.R
\name{pic_calc}
\alias{pic_calc}
\title{Polymorphic Information Content calculation for genind objects}
\usage{
pic_calc(obj, ploidy = 2)
}
\arguments{
\item{obj}{a codominant genind object}

\item{ploidy}{The ploidy of the data}
}
\value{
a data.frame with PIC values (per locus & mean) for each population
and overall
}
\description{
This function is a simple wrapper of the PIC function from
polysat package. It allows to quickly calculate Polymorphic Information
Content or PIC from a genind object. Check PIC help page under polysat
package for references.
}
\author{
Nikolaos Tourvas
}
