% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pid.R
\name{pid_calc}
\alias{pid_calc}
\title{Calculate Probability of Identity (PID) & PIDsibs}
\usage{
pid_calc(obj)
}
\arguments{
\item{obj}{a diploid genind object}
}
\value{
a list with four elements reporting
PID by locus,
PID combinded for all loci,
PIDsibs by locus,
PIDsibs combined for all loci
}
\description{
This function calculates PID & PIDsibs by locus, as well as the
combined statistic for all loci. The current implementation allows only for
diploid data.
}
\examples{
\dontrun{
data(nancycats) # load the "nancycats" genind obj from package adegenet
pid <- pid_calc(obj = nancycats)
}
}
\author{
Nikolaos Tourvas
}
