% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pid.R
\name{pid_permute}
\alias{pid_permute}
\title{Calculate Probability of Identity (PID) & PIDsibs for different
combinations of loci}
\usage{
pid_permute(obj, nrep = 1000)
}
\arguments{
\item{obj}{a diploid genind object}

\item{nrep}{the number of permutations to perform}
}
\value{
a list with three elements, reporting
the raw data.frame from the permutations performed,
the median values of PID & PIDsibs from 1 to the total number of loci,
a boxplot plot
}
\description{
Calculate Probability of Identity (PID) & PIDsibs for different
combinations of loci
}
\examples{
\dontrun{
data(nancycats) # load the "nancycats" genind obj from package adegenet
pid_perm <- pid_permute(obj = nancycats)
}

}
\author{
Nikolaos Tourvas
}
