% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summary_by_loc}
\alias{summary_by_loc}
\title{Produce summary statistics by locus}
\usage{
summary_by_loc(obj)
}
\arguments{
\item{obj}{}
}
\value{

}
\description{
Produce summary statistics by locus
}
