% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summary_by_pop}
\alias{summary_by_pop}
\title{Experimental function: Produce a summary statistics table}
\usage{
summary_by_pop(obj)
}
\arguments{
\item{obj}{a genind object}
}
\value{
a data.frame with summary statistics
}
\description{
This function allows the user to produce to summary table
similar to the one produced by the software GenAlEx for diploid codominant
data. WARNING: This function is still experimental.
}
\author{
Nikolaos Tourvas
}
