% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genind2genambig.R, R/summary_table.R
\name{genind2genambig}
\alias{genind2genambig}
\alias{summary_by_pop}
\title{Convert a genind object to a genambig object}
\usage{
genind2genambig(obj)

summary_by_pop(obj)
}
\arguments{
\item{obj}{a genind object}
}
\value{
a genambig object

a data.frame with summary statistics
}
\description{
Convert a genind object to a genambig object

Produce a summary statistics table
}
\section{Functions}{
\itemize{
\item \code{genind2genambig}: function allows the user to convert a genind object to a
genambig object for analyses using the package polysat.

\item \code{summary_by_pop}: function allows the user to produce to summary table
similar to the one produced by the software GenAlEx.
}}

\author{
Nikolaos Tourvas

Nikolaos Tourvas
}
