% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genemapper2df.R
\name{genemapper2df}
\alias{genemapper2df}
\title{Combine txt files generated by GeneMapper v4.1}
\usage{
genemapper2df(folder_name, export_name = "genemapper_export.csv",
  export_option = "csv")
}
\arguments{
\item{folder_name}{a folder where all the .txt files
generated by GeneMapper are stored.}

\item{export_name}{the name of the exported csv file.}

\item{export_option}{whether to export a csv file (option "csv")
or a data.frame in the R Environment (option "R").}
}
\value{
a single csv file or a data.frame
}
\description{
This function allows the user to quickly convert multiple txt
files of SSR genotyping data from GeneMapper v4.1 to a csv or a data.frame.
}
\examples{
\dontrun{

# Export csv file for use in GenAlEx
genemapper2df(folder_name = "...", export_name = "my_data", export_option = "csv")

# Create genind object
df <- genemapper2df(folder_name = "...", export_option = "R")
obj <- loci2genind(as.loci(df))

# For genind objects you will probably want to import population data as well.
# This can be done in the following way.

strata(obj) <- strata_df    #provide a data.frame with population strata (e.g. Year, Plot)
setPop(obj) <- ~Year/Plot    #set the appropriate population hierachy for the analysis
## Note that you can change population hierarchy between analyses using the previous command.
}
}
\author{
Nikolaos Tourvas
}
