% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summary_by_pop}
\alias{summary_by_pop}
\title{Produce a summary statistics table}
\usage{
summary_by_pop(obj)
}
\arguments{
\item{obj}{a genind object}
}
\value{
a data.frame with summary statistics
}
\description{
Produce a summary statistics table
}
\section{Functions}{
\itemize{
\item \code{summary_by_pop}: function allows the user to produce to summary table
similar to the one produced by the software GenAlEx.
}}

\author{
Nikolaos Tourvas
}
